requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.4.4
Flask-Migrate==3.1.0
Flask-Cors==3.1.1
python-dotenv==0.21.0
Flask-Login==0.6.2
Flask-WTF==1.0.1
requests==2.28.2
pytest==7.2.0
Flask-Testing==0.8.1
pymysql==1.0.2
